v0.4.1: Code smell audit & fixes
Security:
- Model::findBy/findById: SQL injection fix (column name whitelist + backtick quoting)
- Controller::requireAuth: Cast user_id to int for type safety
- SettingsController/ChatController: CSRF token POST fallback for AJAX endpoints
- DiscoverController/ChatController/SettingsController: json_decode validation
- AuthController/SettingsController: Enforce password max length (128 chars)

Performance:
- MatchModel::getByUserId: Replace 3 correlated subqueries with efficient JOINs (N+1 fix)
- Profile::recalcCompleteness: Combine 2 COUNT queries into single query
- Discovery::getStack: Extract haversine() helper, deduplicate formula (3 instances -> 1)

Reliability:
- Discovery::getStack: Add parameter count assertion before query execution
- Photo::delete: Handle unlink failure with error_log
- AdminController: Show error flash on invalid user/status
- public/js/app.js: Fix chat polling memory leak (clearInterval on beforeunload)

+++ app/core/Controller.php
@@ -38,7 +38,7 @@ abstract class Controller
         }
 
         return [
-            'id'    => Session::get('user_id'),
+            'id'    => (int)Session::get('user_id'),
             'email' => Session::get('user_email'),
         ];
     }
